IOS-2568 Move logout to AppCoordinator
