+++ Tangem/Modules/Welcome/WelcomeRoutable.swift
@@ -15,5 +15,4 @@ protocol WelcomeRoutable: AnyObject {
     func openShop()
     func openOnboarding(with input: OnboardingInput)
     func openMain(with cardModel: CardViewModel)
-    func dismiss()
 }
