+++ Tangem/SceneDelegate.swift
@@ -32,18 +32,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         appCoordinator.scene(scene, continue: userActivity)
     }
 
-    func sceneDidBecomeActive(_ scene: UIScene) {
-        // Called when the scene has moved from an inactive state to an active state.
-        // Use this method to restart any tasks that were paused (or not yet started) when the scene was inactive.
-        appCoordinator.sceneDidBecomeActive(scene)
-    }
-
-    func sceneWillResignActive(_ scene: UIScene) {
-        // Called when the scene will move from an active state to an inactive state.
-        // This may occur due to temporary interruptions (ex. an incoming phone call).
-        appCoordinator.sceneWillResignActive(scene)
-    }
-
     func scene(_ scene: UIScene, openURLContexts URLContexts: Set<UIOpenURLContext>) {
         appCoordinator.scene(scene, openURLContexts: URLContexts)
     }
